+++ tests/testutils.py
@@ -41,6 +41,22 @@ def cleanup_apps():
         for app in apps:
             firebase_admin.delete_app(app)
 
+def run_without_project_id(func):
+    env_vars = ['GCLOUD_PROJECT', 'GOOGLE_CLOUD_PROJECT']
+    env_values = []
+    for env_var in env_vars:
+        gcloud_project = os.environ.get(env_var)
+        if gcloud_project:
+            del os.environ[env_var]
+        env_values.append(gcloud_project)
+    try:
+        func()
+    finally:
+        for idx, env_var in enumerate(env_vars):
+            gcloud_project = env_values[idx]
+            if gcloud_project:
+                os.environ[env_var] = gcloud_project
+
 
 class MockResponse(transport.Response):
     def __init__(self, status, response):
@@ -70,9 +86,23 @@ class MockRequest(transport.Request):
 
     def __init__(self, status, response):
         self.response = MockResponse(status, response)
+        self.log = []
 
     def __call__(self, *args, **kwargs):
+        self.log.append((args, kwargs))
         return self.response
+
+
+class MockFailedRequest(transport.Request):
+    """A mock HTTP request that fails by raising an exception."""
+
+    def __init__(self, error):
+        self.error = error
+        self.log = []
+
+    def __call__(self, *args, **kwargs):
+        self.log.append((args, kwargs))
+        raise self.error
 
 
 class MockGoogleCredential(credentials.Credentials):